third_party/blink/web_tests/platform/generic/fast/dom/StyleSheet/gc-rule-children-wrappers-expected.txt

Test that custom properties on rule style declarations are not lost after GC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.styleSheets[0].cssRules[0].type is CSSRule.IMPORT_RULE
PASS document.styleSheets[0].cssRules[0].media.foo is 'bar'
PASS document.styleSheets[0].cssRules[0].type is CSSRule.IMPORT_RULE
PASS document.styleSheets[0].cssRules[0].styleSheet.foo is 'bar'
PASS document.styleSheets[0].cssRules[1].type is CSSRule.STYLE_RULE
PASS document.styleSheets[0].cssRules[1].style.foo is 'bar'
PASS document.styleSheets[0].cssRules[2].type is CSSRule.FONT_FACE_RULE
PASS document.styleSheets[0].cssRules[2].style.foo is 'bar'
PASS document.styleSheets[0].cssRules[3].type is CSSRule.MEDIA_RULE
PASS document.styleSheets[0].cssRules[3].cssRules.foo is 'bar'
PASS document.styleSheets[0].cssRules[3].type is CSSRule.MEDIA_RULE
PASS document.styleSheets[0].cssRules[3].media.foo is 'bar'
PASS document.styleSheets[0].cssRules[4].type is CSSRule.KEYFRAMES_RULE
PASS document.styleSheets[0].cssRules[4].cssRules.foo is 'bar'
PASS document.styleSheets[0].cssRules[4].cssRules[0].type is CSSRule.KEYFRAME_RULE
PASS document.styleSheets[0].cssRules[4].cssRules[0].style.foo is 'bar'
PASS document.styleSheets[0].cssRules[5].type is CSSRule.PAGE_RULE
PASS document.styleSheets[0].cssRules[5].style.foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
